requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
opencv-python>=4.5.3.56
numpy>=1.21.2
torch>=1.9.0
transformers>=4.11.3
sentence-transformers>=2.1.0
ffmpeg-python>=0.2.0
psycopg2-binary>=2.9.5
pydantic>=1.8.2
python-dotenv>=0.19.0
streamlit>=1.22.0
requests>=2.28.2
google-generativeai>=0.3.0
